requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
contextvars==2.4
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.17
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==1.4.24
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.5
nodeenv==1.4.0
numpy==1.19.0
oauthlib==3.1.0
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pluggy==0.13.1
praw==7.1.0
prawcore==1.4.0
pre-commit==2.6.0
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
textblob==0.15.3
toml==0.10.1
tox==3.17.1
tqdm==4.48.0
translate==3.5.0
tweepy==3.9.0
typed-ast==1.4.1
update-checker==0.17
urllib3==1.25.9
virtualenv==20.0.27
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
